fix(remote/amqio): #743 r3 — extract replyRouterFor so the test exercises the shipped adapter
The real-router test had its OWN copy of the ErrPeerRootUnreachable ->
TransientRouteError translation. Deleting the translation from the PRODUCTION
adapter at cmd/amq-remote/main.go left all tests green — the test proved the
pattern is sound, not that the product uses it. A duplicated implementation
in a test is not coverage, it is a second implementation that can agree with
a broken first one.

Fix: extracted replyRouterFor(root) as a named function in cmd/amq-remote.
serve() calls carrier.SetReplyRouter(replyRouterFor(c.root)). The test
TestReplyRouterForTransientPeerAbsent (in cmd/amq-remote, not amqio) uses
replyRouterFor(...) and STOPS defining its own closure.
MUTATION: deleted the translation from replyRouterFor -> test FAILED with
'TICK1: message should stay in new (transient), but ImportOnce handled 1'
(message was DLQ'd — the original B1 defect).

make lint: 0 issues. All tests pass under -race.
